llm_doc_query_final (2)/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
requests
PyMuPDF
PyPDF2
python-docx
sentence-transformers
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu